Dump Roblox Chat Script
When creating a new world, add this script in the ServerScriptService. when the server starts, this script will run automatically. What this script does is logs player chats with their name, timestamp, message id, and player message. Then, when the last player leaves, the server will shut down, and when that happens, it will take that data and send it to the S3 Bucket to be stored.